Replace exception handling middleware by exception handler
