--- dotnet/practice-1/Store Management/WebApi/Middleware/ExceptionHandlingMiddleware.cs
@@ -1,88 +1,0 @@
-﻿using Application.Exceptions;
-using Microsoft.AspNetCore.Mvc;
-
-namespace WebApi.Middleware
-{
-    public class ExceptionHandlingMiddleware
-    {
-        private readonly RequestDelegate _next;
-        private readonly ILogger<ExceptionHandlingMiddleware> _logger;
-
-        public ExceptionHandlingMiddleware(
-            RequestDelegate next,
-            ILogger<ExceptionHandlingMiddleware> logger)
-        {
-            _next = next;
-            _logger = logger;
-        }
-
-        public async Task InvokeAsync(HttpContext context)
-        {
-            try
-            {
-                await _next(context);
-            }
-            catch (Exception exception)
-            {
-                if (exception is ValidationException validationException)
-                {
-                    _logger.LogError(
-                        exception,
-                        "Exception occurred: {Message} {@Errors} {@Exception}",
-                        exception.Message,
-                        validationException.Errors,
-                        validationException);
-                }
-                else
-                {
-                    _logger.LogError(exception, "Exception occurred: {Message}", exception.Message);
-                }
-
-                var exceptionDetails = GetExceptionDetails(exception);
-
-                var problemDetails = new ProblemDetails
-                {
-                    Status = exceptionDetails.Status,
-                    Type = exceptionDetails.Type,
-                    Title = exceptionDetails.Title,
-                    Detail = exceptionDetails.Detail,
-                };
-
-                if (exceptionDetails.Errors is not null)
-                {
-                    problemDetails.Extensions["errors"] = exceptionDetails.Errors;
-                }
-
-                context.Response.StatusCode = exceptionDetails.Status;
-
-                await context.Response.WriteAsJsonAsync(problemDetails);
-            }
-        }
-
-        private static ExceptionDetails GetExceptionDetails(Exception exception)
-        {
-            return exception switch
-            {
-                ValidationException validationException => new ExceptionDetails(
-                    StatusCodes.Status400BadRequest,
-                    "ValidationFailure",
-                    "Validation error",
-                    "One or more validation errors has occurred",
-                    validationException.Errors),
-                _ => new ExceptionDetails(
-                    StatusCodes.Status500InternalServerError,
-                    "ServerError",
-                    "Server error",
-                    "An unexpected error has occurred",
-                    null)
-            };
-        }
-
-        internal record ExceptionDetails(
-            int Status,
-            string Type,
-            string Title,
-            string Detail,
-            IEnumerable<object>? Errors);
-    }
-}
